refactor: reorganize mock backend files to group by business domain not by file type
