+++ src/features/calendar/api/get-case-folders.ts
@@ -1,5 +1,5 @@
 import { mockRequest } from "../../../lib/mock-request";
-import { CaseFolderController } from "../../../services/mock-backend/controller/case-folder-controller";
+import { CaseFolderController } from "../../../services/mock-backend/case-folder/case-folder-controller";
 
 const mockApi = async () => {
 	const request = mockRequest.get(`/calendar`);
