+++ src/features/case-folder/api/delete-case-file.ts
@@ -1,5 +1,5 @@
 import { mockRequest } from "../../../lib/mock-request";
-import { CaseFileController } from "../../../services/mock-backend/controller/case-file-controller";
+import { CaseFileController } from "../../../services/mock-backend/case-file/case-file-controller";
 
 const mockApi = async (folderId: string, fileId: string) => {
 	const request = mockRequest.delete(`/case/${folderId}/${fileId}`);
