+++ src/features/case-folder/api/get-client.ts
@@ -1,5 +1,5 @@
 import { mockRequest } from "../../../lib/mock-request";
-import { ClientController } from "../../../services/mock-backend/controller/client-controller";
+import { ClientController } from "../../../services/mock-backend/case-client/client-controller";
 
 const mockApi = async (folderId: string) => {
 	const request = mockRequest.get(`/case/${folderId}`);
