+++ src/features/case-folder/api/update-case-file-deadline.ts
@@ -1,5 +1,5 @@
 import { mockRequest } from "../../../lib/mock-request";
-import { CaseFileController } from "../../../services/mock-backend/controller/case-file-controller";
+import { CaseFileController } from "../../../services/mock-backend/case-file/case-file-controller";
 
 export type UpdateCaseFileDeadlineDTO = number;
