+++ src/features/create-case-folder/api/create-client.ts
@@ -1,5 +1,5 @@
 import { mockRequest } from "../../../lib/mock-request";
-import { ClientController } from "../../../services/mock-backend/controller/client-controller";
+import { ClientController } from "../../../services/mock-backend/case-client/client-controller";
 import { SexOptions } from "../../../utils/types";
 
 export interface CreateClientDTO {
