+++ src/services/mock-backend/case-client/client-controller.ts
@@ -1,4 +1,4 @@
-import { ClientService } from "../service/client-service";
+import { ClientService } from "./client-service";
 
 export class ClientController {
 	static async createClient(request: Request) {
