+++ src/services/mock-backend/case-client/client-service.ts
@@ -1,5 +1,5 @@
 import { CreateClientDTO } from "../../../features/create-case-folder/api/create-client";
-import { ClientDAO } from "../dao/client-dao";
+import { ClientDAO } from "./client-dao";
 
 export class ClientService {
 	static async createClient(client: CreateClientDTO) {
