+++ src/services/mock-backend/case-file/case-file-dao.ts
@@ -2,7 +2,7 @@ import { DOC_STATUS } from "../../../utils/constants/document-status";
 import { CaseFileObj, DocumentStatus } from "../../../utils/types";
 import { CaseFiles } from "../../mock-sql/schemas";
 import { MockSqlTables } from "../../mock-sql/tables";
-import { DAO } from "./dao";
+import { DAO } from "../dao";
 
 export class CaseFileDAO extends DAO {
 	private static CASE_FILE_STORAGE_KEY = MockSqlTables.table.CASE_FILES;
