+++ src/services/mock-backend/case-folder/case-folder-service.ts
@@ -1,9 +1,9 @@
 import { UpdateCaseFolderDeadlineDTO } from "../../../features/dashboard/api/update-deadline";
 import { Firebase } from "../../cloud-storage/firebase";
-import { CaseFileDAO } from "../dao/case-file-dao";
-import { CaseFolderDAO } from "../dao/case-folder-dao";
-import { ClientDAO } from "../dao/client-dao";
-import { FolderLabelDAO } from "../dao/folder-label-dao";
+import { ClientDAO } from "../case-client/client-dao";
+import { CaseFileDAO } from "../case-file/case-file-dao";
+import { FolderLabelDAO } from "../case-folder-label/folder-label-dao";
+import { CaseFolderDAO } from "./case-folder-dao";
 
 export class CaseFolderService {
 	static async getAllCaseFoldersByUserId(userId: string) {
