+++ src/services/mock-backend/user/user-controller.ts
@@ -1,6 +1,6 @@
 import { RegisterCredentialsDTO } from "../../../features/register/api/register";
 import { SignInCredentialsDTO } from "../../../features/sign-in/api/sign-in";
-import { UserService } from "../service/user-service";
+import { UserService } from "./user-service";
 
 export class UserController {
 	static async verifyUser(request: Request) {
